+++ app/models/custom_ability_character.rb
@@ -3,10 +3,7 @@
 # Custom Ability Exalts, for more flexibility than is allowed to Solars
 class CustomAbilityCharacter < Character
   include AbilityExalt
+  include CustomExalt
 
   attribute :exalt_type, :string, default: 'Ability Exalt'
-
-  def custom_exalt?
-    true
-  end
 end