+++ app/models/custom_attribute_character.rb
@@ -3,10 +3,7 @@
 # Custom Attribute Exalts, for more flexibility than is allowed to Solars
 class CustomAttributeCharacter < Character
   include AttributeExalt
+  include CustomExalt
 
   attribute :exalt_type, :string, default: 'Attribute Exalt'
-
-  def custom_exalt?
-    true
-  end
 end