+++ app/models/custom_essence_character.rb
@@ -3,12 +3,9 @@
 # Custom Essence-based Exalts, like 2e Infernals
 class CustomEssenceCharacter < Character
   include Exalt
+  include CustomExalt
 
   attribute :exalt_type, :string, default: 'Essence Exalt'
 
   alias_attribute :charms, :essence_charms
-
-  def custom_exalt?
-    true
-  end
 end