+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - The weapon editor has been rebuilt from the ground up as a pop-up.
 - Weapons can now have arbitrary bonuses added to accuracy, damage, defense, and/or overwhelming
 - Weapons can also have short notes or descriptions added
-- Added Essence to the list of attributes for attack rolls
+- Custom exalts now start with mote pools of 10 instead of 0
 - Added an option to delete your account and all associated data
 - Bug fixes (Thanks Vejuz)